fix: resolve ~86 E2E test failures for PG18 compatibility
- Fix name type (Oid 19) to text cast in resolve_rangevar_schema
- Return Option<String> from resolve_rangevar_schema for CTE/subquery names
- Strip trailing semicolons from pg_get_viewdef() (PG18 behavior)
- Handle matview (relkind=m) and foreign table (relkind=f) in classify
- Extract classify_relkind and strip_view_definition_suffix as pure functions
- Update CHECK constraint for source_type to include MATVIEW, FOREIGN_TABLE
- Update 7 tests expecting rejection of now-supported auto-rewrite features
- Fix INT4/INT8 type mismatch in DDL event test array_length query

4 remaining failures are pre-existing (CDC partitions, CTE multi-ref,
recursive CTE row_data, CASE-in-WHERE deparsing).

+++ tests/e2e_ddl_event_tests.rs
@@ -208,7 +208,7 @@ async fn test_function_change_marks_st_for_reinit() {
     // Verify functions_used was populated
     let func_count: i64 = db
         .query_scalar(
-            "SELECT coalesce(array_length(functions_used, 1), 0) \
+            "SELECT coalesce(array_length(functions_used, 1), 0)::bigint \
              FROM pgstream.pgs_stream_tables WHERE pgs_name = 'evt_func_st'",
         )
         .await;
